# openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "AtomCV API",
    "description" : "Errors carry a translatable `code`, the `params` its message needs, and the `resolutions` the user can act on. The server never sends display text — see the error catalogue in the architecture document.",
    "license" : {
      "name" : "MIT"
    },
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/api/v1/customizations" : {
      "get" : {
        "tags" : [ "Templates" ],
        "summary" : "The settings this profile kept",
        "description" : "Oldest first, which is the order somebody made them in.",
        "operationId" : "listCustomizations",
        "responses" : {
          "404" : {
            "description" : "RESOURCE_NOT_FOUND",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200" : {
            "description" : "Every saved set",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CustomizationResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Templates" ],
        "summary" : "Keep a set of appearance settings under a name",
        "description" : "Bolum 33.2's Layer B, saved. The profile's own\n`preferences.appearance` is still the working set and is\nwhat a generation uses when it names nothing; this is for\nthe person who keeps a dense set for a long CV and a\nroomier one for a short one.\n\nEvery value is bounded by Bolum 33.2's ranges, and the\nranges are why a bad page is not reachable from here.\n\nAt most twenty per profile, and names are unique within\none.",
        "operationId" : "createCustomization",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CustomizationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomizationResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "VALIDATION_FAILED — an unknown template, a name already used, or one set too many",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404" : {
            "description" : "RESOURCE_NOT_FOUND",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/customizations/{id}" : {
      "delete" : {
        "tags" : [ "Templates" ],
        "summary" : "Forget one",
        "description" : "A generation already made with it is unaffected: Bolum 14.5 writes the settings themselves into the selection snapshot rather than an id, so a document can always be re-rendered exactly as it was sent.",
        "operationId" : "deleteCustomization",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "RESOURCE_NOT_FOUND",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          },
          "204" : {
            "description" : "Deleted"
          }
        }
      },
      "patch" : {
        "tags" : [ "Templates" ],
        "summary" : "Rename or re-set one",
        "description" : "The settings are replaced whole. Bolum 33.2's parameters are read together by the renderer, and a half-applied geometry is a page nobody asked for. Sending only a name renames it and leaves the settings.",
        "operationId" : "patchCustomization",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CustomizationPatch"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "404" : {
            "description" : "RESOURCE_NOT_FOUND",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200" : {
            "description" : "The set as it now stands",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomizationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/templates" : {
      "get" : {
        "tags" : [ "Templates" ],
        "summary" : "The templates a CV can be rendered with",
        "description" : "The registry's own list, with the measured capacity of                     each: Bolum 33.5 describes the three by how much they hold,                     and a chooser showing three names and no density asks                     somebody to pick blind.\n\nNo display name and no description — those are sentences,                     and Bolum 35.4's rule is that the server sends a key and                     the client writes the sentence. The id is the key.\n\nThe same list `capabilities.allowedTemplates` publishes,                     which is what a client reads to know which of these it may                     offer. This one says what each of them is.",
        "operationId" : "listTemplates",
        "responses" : {
          "404" : {
            "description" : "RESOURCE_NOT_FOUND",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200" : {
            "description" : "Every template in the registry",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TemplateSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ApiError" : {
        "type" : "object",
        "description" : "RFC 7807 problem detail with a translatable code",
        "properties" : {
          "type" : {
            "type" : "string",
            "description" : "Relative reference identifying the problem type",
            "example" : "/errors/conflicting-preferences"
          },
          "title" : {
            "type" : "string",
            "description" : "Developer-facing English, stable across occurrences. Never displayed.",
            "example" : "Conflicting preferences"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "HTTP status",
            "example" : 409
          },
          "instance" : {
            "type" : "string",
            "description" : "The path that produced it",
            "example" : "/api/v1/generations"
          },
          "code" : {
            "type" : "string",
            "description" : "Translation key: the client resolves errors.{CODE}",
            "enum" : [ "INSUFFICIENT_PROFILE", "UNPARSEABLE_JOB_DESCRIPTION", "CONFLICTING_PREFERENCES", "FEATURE_REQUIRES_ACCOUNT", "QUOTA_EXCEEDED", "ALL_PROVIDERS_UNAVAILABLE", "COMPILATION_FAILED", "PAGE_LIMIT_EXCEEDED", "REWRITE_VALIDATION_FAILED", "COVER_LETTER_REJECTED", "EMBEDDING_UNAVAILABLE", "GENERATION_PAUSED", "UNSUPPORTED_DOCUMENT", "DOCUMENT_TOO_LARGE", "PDF_NOT_TEXT_BASED", "PDF_ENCRYPTED", "EXTRACTION_EMPTY", "EXTRACTION_TIMEOUT", "LANGUAGE_UNDETECTED", "TRANSLATION_FAILED", "PROFILE_QUOTA_EXCEEDED", "ANONYMOUS_SESSION_EXPIRED", "ATOM_LIMIT_EXCEEDED", "NO_ANONYMOUS_PROFILE", "PROFILE_ALREADY_EXISTS", "GENERATION_ARTIFACT_EXPIRED", "GENERATION_SUPERSEDED", "EDIT_NOT_UNDERSTOOD", "CSRF_TOKEN_INVALID", "AUTHENTICATION_REQUIRED", "OAUTH_FAILED", "MAGIC_LINK_INVALID", "RATE_LIMITED", "CHALLENGE_FAILED", "RESOURCE_NOT_FOUND", "VERSION_CONFLICT", "PRECONDITION_REQUIRED", "VALIDATION_FAILED", "INTERNAL_ERROR", "METHOD_NOT_ALLOWED", "NOT_ACCEPTABLE", "UNSUPPORTED_MEDIA_TYPE" ]
          },
          "params" : {
            "type" : "object",
            "additionalProperties" : { },
            "description" : "Values the translated message interpolates. Keys and types are fixed per code; the server refuses to publish anything undeclared.",
            "example" : {
              "pinnedPages" : 2.3,
              "maxPages" : 1
            }
          },
          "resolutions" : {
            "type" : "array",
            "description" : "What the user can do about it. The server owns this list.",
            "items" : {
              "$ref" : "#/components/schemas/Resolution"
            }
          }
        },
        "required" : [ "code", "status" ]
      },
      "Resolution" : {
        "type" : "object",
        "properties" : {
          "action" : {
            "type" : "string",
            "enum" : [ "increase_page_limit", "replace_profile", "keep_existing_profile", "review_pins", "keep_top_pinned", "sign_up", "paste_full_posting", "continue_as_general_cv", "continue_anyway", "switch_to_manual_form", "complete_profile", "retry" ]
          },
          "params" : {
            "type" : "object",
            "additionalProperties" : { }
          }
        }
      },
...
      "CustomizationRequest" : {
        "type" : "object",
        "description" : "Appearance settings to keep",
        "properties" : {
          "name" : {
            "type" : "string",
            "description" : "What to call it; unique within the profile",
            "example" : "Compact, one page",
            "maxLength" : 60,
            "minLength" : 0
          },
          "baseTemplateId" : {
            "type" : "string",
            "description" : "Which template it is built on",
            "example" : "classic",
            "maxLength" : 40,
            "minLength" : 0
          },
          "fontSizePt" : {
            "type" : "number",
            "format" : "double",
            "description" : "9 to 12 (Bolum 33.2)",
            "example" : 10.5,
            "maximum" : 12.0,
            "minimum" : 9.0
          },
          "marginInches" : {
            "type" : "number",
            "format" : "double",
            "description" : "0.4 to 1.0 inches",
            "example" : 0.5,
            "maximum" : 1.0,
            "minimum" : 0.4
          },
          "lineSpacing" : {
            "type" : "number",
            "format" : "double",
            "description" : "0.9 to 1.3",
            "example" : 1.0,
            "maximum" : 1.3,
            "minimum" : 0.9
          },
          "fontFamily" : {
            "type" : "string",
            "description" : "One of the whitelisted families (Bolum 5.5)",
            "example" : "sans"
          },
          "accentColor" : {
            "type" : "string",
            "description" : "Six hex digits, no hash",
            "example" : "1D4ED8",
            "pattern" : "^[0-9A-Fa-f]{6}$"
          }
        },
        "required" : [ "baseTemplateId", "name" ]
      },
      "CustomizationResponse" : {
        "type" : "object",
        "description" : "A saved set of appearance settings",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "baseTemplateId" : {
            "type" : "string"
          },
          "templateVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "fontSizePt" : {
            "type" : "number",
            "format" : "double"
          },
          "marginInches" : {
            "type" : "number",
            "format" : "double"
          },
          "lineSpacing" : {
            "type" : "number",
            "format" : "double"
          },
          "fontFamily" : {
            "type" : "string"
          },
          "accentColor" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "CustomizationPatch" : {
        "type" : "object",
        "description" : "What to change about a saved set",
        "properties" : {
          "name" : {
            "type" : "string",
            "description" : "A new name, or absent to keep the one it has",
            "maxLength" : 60,
            "minLength" : 0
          },
          "baseTemplateId" : {
            "type" : "string",
            "description" : "Present to replace the settings; absent to leave them",
            "maxLength" : 40,
            "minLength" : 0
          },
          "fontSizePt" : {
            "type" : "number",
            "format" : "double",
            "maximum" : 12.0,
            "minimum" : 9.0
          },
          "marginInches" : {
            "type" : "number",
            "format" : "double",
            "maximum" : 1.0,
            "minimum" : 0.4
          },
          "lineSpacing" : {
            "type" : "number",
            "format" : "double",
            "maximum" : 1.3,
            "minimum" : 0.9
          },
          "fontFamily" : {
            "type" : "string"
          },
          "accentColor" : {
            "type" : "string",
            "pattern" : "^[0-9A-Fa-f]{6}$"
          }
        }
      },
...
      "TemplateSummary" : {
        "type" : "object",
        "description" : "A template a CV can be rendered with",
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "The id used everywhere else",
            "example" : "classic"
          },
          "version" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Renderer version; moves when the geometry does",
            "example" : 6
          },
          "pageTextHeightPt" : {
            "type" : "number",
            "format" : "double",
            "description" : "The text height of one page, in points",
            "example" : 648.0
          },
          "baselineSkipPt" : {
            "type" : "number",
            "format" : "double",
            "description" : "One line of body text, in points",
            "example" : 13.6
          },
          "approximateLinesPerPage" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Roughly how many lines of body text fit a page, which is how Bolum 33.5 describes the three",
            "example" : 54
          }
        }
      },
...
    }
  }
}